requirements.txt
streamlit==1.32.0
deep-translator==1.11.4
python-docx==1.1.0
pymupdf==1.23.8
langcodes==3.3.0
langdetect==1.0.9
pandas==2.2.2